requirements.txt
beautifulsoup4
Django~=3.1.14
djangorestframework==3.13.1
django-admin-autocomplete-filter==0.7.1
django-guardian==2.4.0
humanfriendly
jsonfield==3.1.0
requests
